Replace size_t with uint64_t

+++ include/dish2/peripheral/readable_state/introspective_state/raw_introspective_state/KinGroupIDAncestorView.hpp
@@ -13,9 +13,9 @@
 namespace dish2 {
 
 template< typename Spec >
-struct KinGroupIDAncestorView : public uitsl::PodLeafNode< size_t, Spec::NLEV > {
+struct KinGroupIDAncestorView : public uitsl::PodLeafNode< uint64_t, Spec::NLEV > {
   // inherit constructors
-  using parent_t = uitsl::PodLeafNode<size_t, Spec::NLEV>;
+  using parent_t = uitsl::PodLeafNode<uint64_t, Spec::NLEV>;
   using parent_t::parent_t;
 
 };
